Ungate demux from dynamic-graph, and wire the overflow child in both examples
Two deviations from legacy that the demux examples surfaced.

**demux was behind `dynamic-graph`; in legacy it is not.** Legacy's
`nodes/demux.rs` carries no cfg and its `demux` example declares no
`required-features`, so `cargo run --example demux` works on a default
build. Here `demux` / `demux_map` / `demux_it` / `DemuxEvent` sat in the
same `#[cfg(feature = "dynamic-graph")] impl Builder` as `dynamic_group`,
which cost a default-features user the capability outright. Nothing about
demux needs the feature: it adds and removes no nodes — the slot pool is
wired once and only the tick is routed, through the same-cycle mark-dirty
buffer (`marks` / `has_marks`), which the engine already compiles
unconditionally. Split the routing methods into their own ungated `impl
Builder`, ungate `DemuxEvent` / `DemuxMap`, and drop `required-features`
from both example targets. The graph *mutation* surface (`run_dynamic`,
`Extension`, `dynamic_group`) stays gated exactly as before.

The five demux tests move out of `tests/dynamic_graph.rs` (whose
file-level `#![cfg(feature = "dynamic-graph")]` would have kept them off a
default build) into `tests/demux.rs`, so the routing surface is covered by
a plain `cargo test`.

**Both examples dropped their overflow child on the floor.** Legacy wires
`overflow.panic()` into the graph, so exceeding the pool aborts loudly;
here `main` bound `_overflow` and never observed it, leaving unrouted
events silently discarded in the one place a reader looks for how demux
handles them — while the group README lists wiring an overflow path as
demux's headline cost. Both now abort the run through a `for_each`
returning `Err`, the wingfoil spelling of legacy's `.panic()`. Verified by
temporarily setting `CAPACITY = 1`: the run stops on the first unrouted
event with the burst in the error context. The tests still assert the
child never fires at the real capacity.

Book states, tick times and both READMEs' sample output are unchanged.
